project part 6/p6_i/requirements.txt
pyyaml
sklearn-crfsuite
numpy
conlleval